requirements.txt
xarray 
numpy 
matplotlib 
pandas
h5netcdf
scipy
chromadb
requests
python-dotenv
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cu126
groq
seaborn